BugFix: Prevent duplicate rewards on player logout.
- Add game status check at `onPlayerQuit` event handler.

+++ src/main/java/dev/spring93/sumoevent/listeners/PlayerQuitListener.java
@@ -12,6 +12,8 @@ public class PlayerQuitListener implements Listener {
 
     @EventHandler
     public void onPlayerQuit(PlayerQuitEvent event) {
+        if(!gameService.isGameActive()) return;
+
         Player player = event.getPlayer();
         gameService.removePlayerFromQueue(player);
